selinux: Conditionally remove libselinux mask
If the meta-qti-sepolicy layer is included, the build system
is required to pick up the libselinux version provided by
meta-selinux. However, the meta-selinux library is only installed
when selinux is in distro features. This causes incompatibilities
if a build requires libselinux from any source but hasn't yet
included selinux in distro features.

Also remove existing instances of the deprecated base_contains()

Change-Id: Ifca88945f86c552c27617b50fe3475cd05adc8c3