Add missing selinux policy rules to avoid avc denials.
avc denials can be listed from the dmesg logs. When the device is in
permissive mode i.e selinux allows logging of the would be denials by
the selinux framework and simultaneous execution of the operations in
an  inactive selinux framework environment. The kernel hooks are
allowed to be used even though they can't had the device been in
enforcing mode.

Change-Id: I0ecef3e10ca1f3875fad6ca1388439550919b184