selinux: Create /dev/socket after /dev remount
When the sysfs initscript remounts /dev, /dev/socket is deleted.
The leprop service will then recreate the directory and immediately
attempt to use it without attempting to assign the correct selinux
context. Create /dev/socket just before performing a recursive
restorecon call to ensure this is created with the proper label.

Change-Id: I181d56e69220d3c51d831b224fc072d7b302aaf3