selinux: Add pam to DISTRO_FEATURES_BACKFILL
Libpam is now a dependency for selinux, so add pam to
DISTRO_FEATURES_BACKFILL to ensure it is always included on machines
using selinux. A side-effect of the change is that console logins
receive the root:sysadm_r:sysadm_t context by default. Move both
local_login and adb shells to the root:unconfined_r:unconfined_t
context on userdebug builds.

Change-Id: I47d2d3bf4a7411bebcbed44242dfee71beefddb2