selinux: Move selinux-labeldev to tmpfiles
Move the functionality of the selinux-labeldev script to
systemd-tmpfiles. Character files inside /dev are already handled
via udev so there are very few files in /dev that would need to have
SELinux labels assigned. Use systemd-tmpfiles to cover the remaining
files without going through the overhead of launching a systemd
service or performing a full recursive restorecon on /dev.

Change-Id: Ic95bd60c257913cbe85d9b862882980a5af1b902