sepolicy: Fixup userdebug_or_eng usage
The 'kerneltests_modules_t' is defined inside userdebug_or_eng, but
used in a file contexts file on user variant as well. This leads to
a compilation failure, so move the declaration outside of the macro.

optional_policy blocks require some gen_require statement (or
possibly other types of statements) inside the block. When another
block (in this case userdebug_or_eng) strips away all statements
inside the block, it leads to compilation failures. Swap the order
so that userdebug_or_eng will also eliminate the optional_policy
call along with all the wrapped statements.

Change-Id: I92280498227b98e17c8c16d8ff87486f3df8a997